requirements.txt
astroid==3.0.1
autopep8==2.0.4
bcrypt==4.0.1
bidict==0.22.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coverage==7.3.2
dill==0.3.7
setuptools
wheel==0.45.0
exceptiongroup==1.1.3
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Testing==0.8.1
flask-socketio
Flask-Migrate
pandas
numpy
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
h11==0.14.0
idna==3.6
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.2
platformdirs==4.0.0
pluggy==1.3.0
pycodestyle==2.11.1
pylint==3.0.2
pytest==7.4.2
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3.post1
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.23
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.1